requirements.txt
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0
boto3>=1.21.40
cdk-nag>=2.14.11